backend/pyproject.toml
[project]
name = "breeze-backend"
version = "0.1.0"
description = "Backend for Breeze, where AI agents are orchestrated and managed."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "langfuse>=3.14.5",
    "mcp>=1.26.0",
    "ollama>=0.6.1",
    "openai>=2.25.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "slowapi>=0.1.9",
    "tavily-mcp>=0.2.3",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
]
